requirements.txt
asgiref==3.8.1
Django==5.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
mccabe==0.7.0
packaging==24.0
phonenumbers==8.13.38
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
uritemplate==4.1.1
